Sentiment Trend Prediction for the Indian General Election
The popularity and accessibility of social media applications, such as Twitter, have
increased over the past few years at a rapid pace. Users from different parts of the world use it to
express opinions and share thoughts with each other. Politicians all over the world suitably use these
platforms at every opportunity and this has resulted in an increase in support and following for
themselves and their parties. Sentiment analysis has become a key methodology to gain insight from
social networks. The aim of this project is to perform sentiment analysis using a lexicon-based
sentiment analyzer and sentiment trend prediction for a short interval of time for modeling the
sentiment trend towards the top contenders of the Indian General Election 2019 on twitter. Sentiment
Analysis is done on the tweets posted during the campaign period of the Indian General Elections of
2019. This is done using a pre-trained sentiment analyzer called VADER of the NLTK library, which is
written in Python. VADER is optimized for social media data and can yield good results when used
with data from Twitter, Facebook, etc. The dataset used for this project is created using web scraping
modules written in Python.In addition to that, sentiment trend prediction was done for a period of 10
days from the day of result announcement using Linear Regression